+++ apps/pie-storybook/stories/pie-cookie-banner.stories.ts
@@ -1,36 +1,38 @@
 import { html, TemplateResult } from 'lit';
-import { PieCookieBanner, CookieBannerProps } from '@justeattakeaway/pie-cookie-banner';
-import { PieButton } from '@justeattakeaway/pie-button';
-import { PieLink } from '@justeattakeaway/pie-link';
-import { PieModal } from '@justeattakeaway/pie-modal';
-import { PieIconButton } from '@justeattakeaway/pie-icon-button';
-import { PieToggleSwitch } from '@justeattakeaway/pie-toggle-switch';
+import { action } from '@storybook/addon-actions';
+
+/* eslint-disable import/no-duplicates */
+import '@justeattakeaway/pie-cookie-banner';
+import { CookieBannerProps } from '@justeattakeaway/pie-cookie-banner';
+import pieCookieBannerLocales from '@justeattakeaway/pie-cookie-banner/locales';
+/* eslint-enable import/no-duplicates */
 
 import { type StoryMeta } from '../types';
 import { createStory } from '../utilities';
 
-// This prevents storybook from tree shaking the components
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const keptReferences = [
-    PieCookieBanner,
-    PieButton,
-    PieLink,
-    PieModal,
-    PieIconButton,
-    PieToggleSwitch
-];
-
 type CookieBannerStoryMeta = StoryMeta<CookieBannerProps>;
 
 const defaultArgs: CookieBannerProps = {
-
+    hasPrimaryActionsOnly: false,
+    locale: pieCookieBannerLocales.enGB,
 };
 
 const cookieBannerStoryMeta: CookieBannerStoryMeta = {
     title: 'Cookie Banner',
     component: 'pie-cookie-banner',
     argTypes: {
-
+        hasPrimaryActionsOnly: {
+            description: 'When true, sets the variant to "primary" for the button which accepts necessary cookies only.',
+            control: 'boolean',
+        },
+        locale: {
+            options: Object.keys(pieCookieBannerLocales),
+            mapping: pieCookieBannerLocales,
+            description: 'Assigns the data for localising the component strings',
+            defaultValue: {
+                summary: 'enGB',
+            },
+        },
     },
     args: defaultArgs,
     parameters: {
@@ -43,8 +45,23 @@ const cookieBannerStoryMeta: CookieBannerStoryMeta = {
 
 export default cookieBannerStoryMeta;
 
-const BaseStoryTemplate = () : TemplateResult => html`
-        <pie-cookie-banner></pie-cookie-banner>`;
+const necessaryOnlyAction = action('necessary-only');
+const acceptAllAction = action('accept-all');
+const managePrefsAction = action('manage-prefs');
+const prefsSavedAction = action('prefs-saved');
+
+const BaseStoryTemplate = (props: CookieBannerProps) : TemplateResult => {
+    const { hasPrimaryActionsOnly, locale } = props;
+
+    return html`
+        <pie-cookie-banner
+            .locale=${locale}
+            ?hasPrimaryActionsOnly="${hasPrimaryActionsOnly}"
+            @pie-cookie-banner-necessary-only="${necessaryOnlyAction}"
+            @pie-cookie-banner-accept-all="${acceptAllAction}"
+            @pie-cookie-banner-manage-prefs="${managePrefsAction}"
+            @pie-cookie-banner-prefs-saved="${prefsSavedAction}"></pie-cookie-banner>`;
+};
 
 /**
  * Creates a 'page' of scrollable HTML. Useful for testing scroll behaviours in a Story.
@@ -61,15 +78,9 @@ const createScrollablePageHTML = () : TemplateResult => {
         <ul>${items}</ul>`;
 };
 
-// TODO: remove the eslint-disable rule when props are added
-// eslint-disable-next-line no-empty-pattern
-const Template = ({}: CookieBannerProps): TemplateResult => html`
-    <pie-cookie-banner></pie-cookie-banner>
-`;
-
-const ScrollablePageStoryTemplate = () : TemplateResult => html`
-    ${BaseStoryTemplate()}
+const ScrollablePageStoryTemplate = (props: CookieBannerProps) : TemplateResult => html`
+    ${BaseStoryTemplate(props)}
     ${createScrollablePageHTML()}`;
 
-export const Default = createStory<CookieBannerProps>(Template, defaultArgs)();
+export const Default = createStory<CookieBannerProps>(BaseStoryTemplate, defaultArgs)();
 export const ScrollablePage = createStory<CookieBannerProps>(ScrollablePageStoryTemplate, defaultArgs)();