+++ packages/components/pie-icon-button/src/index.ts
@@ -1,42 +1,70 @@
-import { LitElement, html, unsafeCSS } from 'lit';
+import {
+    LitElement, TemplateResult, html, unsafeCSS,
+} from 'lit';
 import { property } from 'lit/decorators.js';
-import { validPropertyValues } from '@justeattakeaway/pie-webc-core';
-
+import { validPropertyValues, defineCustomElement } from '@justeattakeaway/pie-webc-core';
 import styles from './iconButton.scss?inline';
 import {
     IconButtonProps, sizes, variants,
 } from './defs';
+import '@justeattakeaway/pie-spinner';
 
 // Valid values available to consumers
 export * from './defs';
 
 const componentSelector = 'pie-icon-button';
 
+/**
+ * @tagname pie-icon-button
+ */
 export class PieIconButton extends LitElement implements IconButtonProps {
     @property()
     @validPropertyValues(componentSelector, sizes, 'medium')
-    public size: IconButtonProps['size'] = 'medium';
+    public size?: IconButtonProps['size'] = 'medium';
 
     @property()
     @validPropertyValues(componentSelector, variants, 'primary')
-    public variant: IconButtonProps['variant'] = 'primary';
+    public variant?: IconButtonProps['variant'] = 'primary';
 
     @property({ type: Boolean })
-    public disabled = false;
+    public disabled? = false;
+
+    @property({ type: Boolean })
+    public isLoading? = false;
+
+    /**
+     * Template for the loading state
+     *
+     * @private
+     */
+    private renderSpinner (): TemplateResult {
+        const { variant, size } = this;
+        const spinnerSize = size === 'xsmall' ? 's' : 'm';
+        let spinnerVariant = 'brand';
+        if (variant?.includes('secondary')) spinnerVariant = 'secondary';
+        if (variant === 'primary' || variant === 'ghost-inverse') spinnerVariant = 'inverse';
+
+        return html`
+                <pie-spinner
+                    size="${spinnerSize}"
+                    variant="${spinnerVariant}"
+                </pie-spinner>`;
+    }
 
     render () {
         const {
-            disabled, size, variant,
+            disabled, size, variant, isLoading,
         } = this;
 
         // The inline SVG is temporary until we have a proper icon integration
         return html`
             <button
                 class="o-iconBtn"
-                size=${size}
-                variant=${variant}
-                ?disabled=${disabled}>
-                <slot></slot>
+                size="${size}"
+                variant="${variant}"
+                ?disabled="${disabled}"
+                ?isLoading="${isLoading}">
+                ${isLoading ? this.renderSpinner() : html`<slot></slot>`}
             </button>`;
     }
 
@@ -44,7 +72,7 @@ export class PieIconButton extends LitElement implements IconButtonProps {
     static styles = unsafeCSS(styles);
 }
 
-customElements.define(componentSelector, PieIconButton);
+defineCustomElement(componentSelector, PieIconButton);
 
 declare global {
     interface HTMLElementTagNameMap {