+++ packages/components/pie-modal/src/index.ts
@@ -3,13 +3,16 @@ import {
 } from 'lit';
 import { html, unsafeStatic } from 'lit/static-html.js';
 import { property, query } from 'lit/decorators.js';
+import { disableBodyScroll, enableBodyScroll } from 'body-scroll-lock';
+
+import '@justeattakeaway/pie-button';
+import '@justeattakeaway/pie-icon-button';
 import {
-    requiredProperty, RtlMixin, validPropertyValues,
+    requiredProperty, RtlMixin, validPropertyValues, defineCustomElement,
 } from '@justeattakeaway/pie-webc-core';
 import '@justeattakeaway/pie-icons-webc/IconClose';
 import '@justeattakeaway/pie-icons-webc/IconChevronLeft';
 import '@justeattakeaway/pie-icons-webc/IconChevronRight';
-import { disableBodyScroll, enableBodyScroll } from 'body-scroll-lock';
 
 import styles from './modal.scss?inline';
 import {
@@ -33,6 +36,7 @@ export * from './defs';
 const componentSelector = 'pie-modal';
 
 /**
+ * @tagname pie-modal
  * @event {CustomEvent} pie-modal-open - when the modal is opened.
  * @event {CustomEvent} pie-modal-close - when the modal is closed.
  * @event {CustomEvent} pie-modal-back - when the modal back button is clicked.
@@ -452,7 +456,7 @@ export class PieModal extends RtlMixin(LitElement) implements ModalProps {
     };
 }
 
-customElements.define(componentSelector, PieModal);
+defineCustomElement(componentSelector, PieModal);
 
 declare global {
     interface HTMLElementTagNameMap {