+++ packages/components/pie-spinner/test/accessibility/pie-spinner.spec.ts
@@ -1,0 +1,18 @@
+
+import { test, expect } from '@justeattakeaway/pie-webc-testing/src/playwright/fixtures.ts';
+import { PieSpinner, SpinnerProps } from '@/index';
+
+test.describe('PieSpinner - Accessibility tests', () => {
+    test('a11y - should test the PieSpinner component WCAG compliance', async ({ makeAxeBuilder, mount }) => {
+        await mount(
+            PieSpinner,
+            {
+                props: {} as SpinnerProps,
+            },
+        );
+
+        const results = await makeAxeBuilder().analyze();
+
+        expect(results.violations).toEqual([]);
+    });
+});