+++ packages/components/pie-icon-button/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @justeattakeaway/pie-icon-button
 
+## 0.21.0
+
+### Minor Changes
+
+- [Added] - `isLoading` prop to `pie-icon-button` ([#953](https://github.com/justeattakeaway/pie/pull/953)) by [@raoufswe](https://github.com/raoufswe)
+
+  [Added] - `visually-hidden` mixin to `pie-css` and use it in components when needed
+  [Added] - `button-interactive-states` mixin to use in `pie-icon-button` and `pie-button`
+
+### Patch Changes
+
+- Updated dependencies [[`b7bcd63d1`](https://github.com/justeattakeaway/pie/commit/b7bcd63d1d035f89a8b4db8ea5977b9420d6f9d5), [`72b21e8ad`](https://github.com/justeattakeaway/pie/commit/72b21e8ad03439655653edab01acbb1a78a94ecf)]:
+  - @justeattakeaway/pie-spinner@0.2.0
+
+## 0.20.0
+
+### Minor Changes
+
+- [Added] - `inverse` and `ghost-inverse` variants ([#941](https://github.com/justeattakeaway/pie/pull/941)) by [@raoufswe](https://github.com/raoufswe)
+
+  [Fixed] - Icon sizes aren't correct
+
+  [Fixed] - The `outline` variant border is missing
+
+  [Fixed] - The inner ring of focus styles is missing
+
+  [Fixed] - The background of `outline`, `ghost` and `ghost-secondary` variants should be transparent
+
+  [Fixed] - The background of `outline`, `ghost` and `ghost-secondary` variants should be transparent
+
+  [Fixed] - TH disabled state should use the `not-allowed` cursor
+
+## 0.19.0
+
+### Minor Changes
+
+- [Removed] - pie-monorepo specific content from Web Component READMES ([#904](https://github.com/justeattakeaway/pie/pull/904)) by [@LTurns](https://github.com/LTurns)
+
+## 0.18.0
+
+### Minor Changes
+
+- [Added] - @tagname jsdoc comment to top of component class and use new defineCustomElement function to define the components ([#905](https://github.com/justeattakeaway/pie/pull/905)) by [@jamieomaguire](https://github.com/jamieomaguire)
+
 ## 0.17.0
 
 ### Minor Changes