+++ packages/tools/pie-css/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @justeattakeaway/pie-css
 
+## 0.7.0
+
+### Minor Changes
+
+- [Added] - `isLoading` prop to `pie-icon-button` ([#953](https://github.com/justeattakeaway/pie/pull/953)) by [@raoufswe](https://github.com/raoufswe)
+
+  [Added] - `visually-hidden` mixin to `pie-css` and use it in components when needed
+  [Added] - `button-interactive-states` mixin to use in `pie-icon-button` and `pie-button`
+
 ## 0.6.0
 
 ### Minor Changes